requirements/base.txt
# requirements/base.txt
# Core Django
Django==5.0.3
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
django-model-utils==4.4.0
django-redis==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-health-check==3.17.0
django-structlog==7.0.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27

# Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Security
argon2-cffi==23.1.0
cryptography==42.0.2
django-permissions-policy==4.19.0
django-csp==3.7

# Performance & Caching
django-cache-memoize==0.1.10
django-cachalot==2.6.3
django-prometheus==2.3.1

# File Handling & Processing
Pillow==10.2.0
python-magic==0.4.27
openpyxl==3.1.2
pandas==2.2.0
numpy==1.26.4

# Cloudflare R2 Storage
django-storages==1.14.2
boto3==1.34.34  # Compatible with S3/R2 API

# Utils
python-dateutil==2.8.2
pytz==2024.1
pendulum==3.0.0
pydantic==2.6.1
python-slugify==8.0.4
phonenumbers==8.13.30

# Validation & Serialization
marshmallow==3.20.2
pyyaml==6.0.1

# Monitoring & Logging
sentry-sdk==1.40.3
elastic-apm==6.19.0
structlog==24.1.0
python-json-logger==2.0.7